requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
typing-extensions>=4.11.0
